// src/types/character.ts
export interface Character {
  id: number;
  name: string;
  title: string;
  backstory: string;
  class: string;
  race: string;
  skills: string[];
  languages: string[];
  background: string;
  image: string;
  backgroundImage: string;
  thumbnail: string;
  classIcon: string; 
  raceIcon: string;
} 